move classes into 'models/'
