+++ src/index.ts
@@ -1,6 +1,6 @@
-export { default as Arg } from "./Arg";
+export { default as Arg } from "./models/Arg";
 export { default as ArgParser } from "./ArgParser";
-export { default as Command } from "./Command";
-export { default as CommandResultStatus } from "./CommandResultStatus";
-export { default as CommandSet } from "./CommandSet";
-export { default as ParseOption } from "./ParseOption";
+export { default as Command } from "./models/Command";
+export { default as CommandResultStatus } from "./models/CommandResultStatus";
+export { default as CommandSet } from "./models/CommandSet";
+export { default as ParseOption } from "./models/ParseOption";
