+++ src/models/CommandResult.ts
