+++ src/models/CommandResultStatus.ts
