+++ src/models/Prop.ts
